Atm Uygulaması devam ediyor

+++ Users.cs
@@ -6,13 +6,13 @@ namespace Atm.Atm
         private string surname;
         private string idno;
         private string password;
-        private double balance;
+        private int balance;
 
         public string Name { get => name; set => name = value; }
         public string Surname { get => surname; set => surname = value; }
         public string Idno { get => idno; set => idno = value; }
     
         public string Password { get => password; set => password = value; }
-        public double Balance { get => balance; set => balance = value; }
+        public int Balance { get => balance; set => balance = value; }
     }
 }
